Synchronized migrations with definitions + rights migration
There was stuff present in migrations and not present in definitions - I
added that to definitions. There were some changes in definitions that
didn't have anything associated with them in migration - I added that.

Also, since we are using definitions-lib now and the rights actually
work there, might as well commit rights management into the migrations.